feat: add FP8 KV cache and precompiled Docker deployment
Add calibrated per-layer FP8 E4M3 KV caches for LLaDA2 across SDPA/Flex and FlashInfer, including paged writes and existing CUDA Graph integration.

Build native kernels and Torch DLPack bridges in Docker, add persistent compiler caches and startup warmup, and integrate worker and graph cleanup into HTTP shutdown.

Document calibration, deployment, and validation limits. Verified 94 CPU regressions with 2 skips, four-H200 TP/EP eager serving on both attention backends, warmup-only exit, and reuse of 119 compiled artifacts.

+++ python/fluxserve/backend/managers/kvcache/dense.py
@@ -108,7 +108,7 @@ class KVCache:
 
             device = self._raw_data.device
             num_layer, _, batch_size, num_heads, seq_len, hidden_dim = self._raw_data.shape
-            self._data = torch.zeros(num_layer, 2, batch_size, num_heads, self.length, hidden_dim, device=device, dtype=torch.bfloat16)
+            self._data = torch.zeros(num_layer, 2, batch_size, num_heads, self.length, hidden_dim, device=device, dtype=self._raw_data.dtype)
             self._data[:, :, :, :, :seq_len] = self._raw_data
 
     def consolidate(self):
